Revert "Cortex_M backend: Add conv op (#15896)"
This reverts commit 213d24c514d9009a5aab27a0eb074c6cc7f01e65.
