+++ backends/cortex_m/ops/operators.py
@@ -6,10 +6,8 @@
 # LICENSE file in the root directory of this source tree.
 
 from math import prod
-from typing import Sequence
 
 import torch
-import torch.nn.functional as F
 from executorch.backends.cortex_m.passes.passes_utils import (
     requantize_cmsis,
     SHIFT_INT8,
@@ -410,163 +408,3 @@ def transpose_meta(input: torch.Tensor, perm) -> torch.Tensor:
 @impl(lib, "transpose", "CompositeExplicitAutograd")
 def transpose_impl(input: torch.Tensor, perm) -> torch.Tensor:
     return input.permute(tuple(perm)).contiguous()
-
-
-# ===================================================================
-# QUANTIZED CONV2D OPERATION DEFINITION
-# ===================================================================
-
-lib.define(
-    "quantized_conv2d("
-    "Tensor input, "
-    "Tensor weight, "
-    "Tensor? bias, "
-    "int[] stride, "
-    "int[] padding, "
-    "int[] dilation, "
-    "int input_offset, "
-    "int output_offset, "
-    "Tensor requantize_multipliers, "
-    "Tensor requantize_shifts, "
-    "int activation_min, "
-    "int activation_max"
-    ") -> Tensor"
-)
-
-
-lib.define(
-    "quantized_conv2d.out("
-    "Tensor input, "
-    "Tensor weight, "
-    "Tensor? bias, "
-    "int[] stride, "
-    "int[] padding, "
-    "int[] dilation, "
-    "int input_offset, "
-    "int output_offset, "
-    "Tensor requantize_multipliers, "
-    "Tensor requantize_shifts, "
-    "int activation_min, "
-    "int activation_max, "
-    "*, Tensor(a!) out"
-    ") -> Tensor(a!)"
-)
-
-
-def _compute_conv2d_output_shape(
-    input_shape: torch.Size,
-    weight_shape: torch.Size,
-    stride: Sequence[int],
-    padding: Sequence[int],
-    dilation: Sequence[int],
-) -> torch.Size:
-    batch = input_shape[0]
-    in_height = input_shape[2]
-    in_width = input_shape[3]
-    # We store the weights in OHWI layout (out, kernel_h, kernel_w, in)
-    kernel_height = weight_shape[1]
-    kernel_width = weight_shape[2]
-
-    stride_h, stride_w = stride
-    pad_h, pad_w = padding
-    dilation_h, dilation_w = dilation
-
-    out_channels = weight_shape[0]
-    out_height = (
-        in_height + 2 * pad_h - dilation_h * (kernel_height - 1) - 1
-    ) // stride_h + 1
-    out_width = (
-        in_width + 2 * pad_w - dilation_w * (kernel_width - 1) - 1
-    ) // stride_w + 1
-    return torch.Size([batch, out_channels, out_height, out_width])
-
-
-@register_fake("cortex_m::quantized_conv2d")
-def quantized_conv2d_meta(
-    input: torch.Tensor,
-    weight: torch.Tensor,
-    bias: torch.Tensor | None,
-    stride: Sequence[int],
-    padding: Sequence[int],
-    dilation: Sequence[int],
-    input_offset: int,
-    output_offset: int,
-    requantize_multipliers: torch.Tensor,
-    requantize_shifts: torch.Tensor,
-    activation_min: int,
-    activation_max: int,
-) -> torch.Tensor:
-    stride_vals = list(stride)
-    padding_vals = list(padding)
-    dilation_vals = list(dilation)
-    output_shape = _compute_conv2d_output_shape(
-        input.shape, weight.shape, stride_vals, padding_vals, dilation_vals
-    )
-    return torch.empty(
-        output_shape,
-        dtype=torch.int8,
-        device=input.device,
-        memory_format=torch.channels_last,
-    )
-
-
-@impl(lib, "quantized_conv2d", "CompositeExplicitAutograd")
-def quantized_conv2d_impl(
-    input: torch.Tensor,
-    weight: torch.Tensor,
-    bias: torch.Tensor | None,
-    stride: Sequence[int],
-    padding: Sequence[int],
-    dilation: Sequence[int],
-    input_offset: int,
-    output_offset: int,
-    requantize_multipliers: torch.Tensor,
-    requantize_shifts: torch.Tensor,
-    activation_min: int,
-    activation_max: int,
-) -> torch.Tensor:
-    if input.dim() != 4 or weight.dim() != 4:
-        raise RuntimeError("quantized_conv2d expects 4D input and weight tensors")
-    # Convert to int32 for accumulation and apply offsets
-    input_int32 = input.to(torch.int32) + int(input_offset)
-    weight_int32 = weight.to(torch.int32)
-
-    if bias is None:
-        bias_int32 = torch.zeros(
-            weight.shape[0], dtype=torch.int32, device=input.device
-        )
-    else:
-        bias_int32 = bias.to(torch.int32)
-
-    input_channels = input.shape[1]
-    kernel_input_channels = weight.shape[3]
-    groups = input_channels // kernel_input_channels
-
-    # Convert weights back to OIHW layout expected by torch.nn.functional.conv2d
-    weight_oi_hw = weight_int32.permute(0, 3, 1, 2).contiguous()
-
-    conv_acc = F.conv2d(
-        input_int32,
-        weight_oi_hw,
-        bias_int32,
-        stride=tuple(stride),
-        padding=tuple(padding),
-        dilation=tuple(dilation),
-        groups=groups,
-    )
-
-    result_channels = []
-    for output_channel_i in range(conv_acc.shape[1]):
-        result_channel = requantize_cmsis(
-            conv_acc[:, output_channel_i, :, :],
-            int(requantize_multipliers[output_channel_i]),
-            int(requantize_shifts[output_channel_i]),
-        )
-        result_channels.append(result_channel)
-
-    result = torch.stack(result_channels, dim=1)
-
-    result += output_offset
-    result = torch.clamp(result, activation_min, activation_max)
-
-    return result.to(torch.int8)
