+++ backends/cortex_m/ops/operators.yaml
@@ -52,9 +52,3 @@
   kernels:
     - arg_meta: null
       kernel_name: cortex_m::transpose_out
-
-- func: cortex_m::quantized_conv2d.out(Tensor input, Tensor weight, Tensor? bias, int[] stride, int[] padding, int[] dilation, int input_offset, int output_offset, Tensor requantize_multipliers, Tensor requantize_shifts, int activation_min, int activation_max, *, Tensor(a!) out) -> Tensor(a!)
-  variants: function
-  kernels:
-    - arg_meta: null
-      kernel_name: cortex_m::quantized_conv2d_out
