+++ backends/cortex_m/passes/__init__.py
@@ -3,7 +3,7 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .convert_to_cortex_m_pass import ConvertToCortexMPass  # noqa
+from .quantized_linear_fusion_pass import QuantizedLinearFusionPass  # noqa
 from .quantized_op_fusion_pass import QuantizedOpFusionPass  # noqa
 from .replace_quant_nodes_pass import ReplaceQuantNodesPass  # noqa
 from .cortex_m_pass_manager import CortexMPassManager  # noqa  # usort: skip
