+++ backends/cortex_m/passes/cortex_m_pass_manager.py
@@ -4,34 +4,30 @@
 # LICENSE file in the root directory of this source tree.
 
 
-import inspect
-
 from executorch.backends.arm._passes import (
     FoldAndAnnotateQParamsPass,
     ScalarsToAttributePass,
 )
 from executorch.backends.cortex_m.passes import (
-    ConvertToCortexMPass,
+    QuantizedLinearFusionPass,
     QuantizedOpFusionPass,
     ReplaceQuantNodesPass,
 )
 from executorch.backends.transforms.replace_scalar_with_tensor import (
     ReplaceScalarWithTensorArgPass,
 )
+from executorch.backends.xnnpack._passes import XNNPACKPassManager
 from executorch.exir.pass_base import ExportPass
-from executorch.exir.pass_manager import PassManager
-from executorch.exir.program._program import _transform
-from torch.export import ExportedProgram
 
 
-class CortexMPassManager(PassManager):
+class CortexMPassManager(XNNPACKPassManager):
 
     pass_list: list[ExportPass] = [
         FoldAndAnnotateQParamsPass,
         ReplaceScalarWithTensorArgPass,
         ReplaceQuantNodesPass,
         QuantizedOpFusionPass,
-        ConvertToCortexMPass,
+        QuantizedLinearFusionPass,
     ]
 
     pass_list_transform_for_annotation: list[ExportPass] = [
@@ -40,29 +36,10 @@ class CortexMPassManager(PassManager):
     ]
 
     def __init__(self, exported_program, passes=None):
-        self.exported_program = exported_program
-        if passes is not None:
-            self.passes = passes
-        else:
-            self.passes = self.pass_list
+        super().__init__(exported_program, passes or self.pass_list)
 
     def transform_for_annotation(self, model):
         passes = self.pass_list_transform_for_annotation
         for p in passes:
             model = p().call(model).graph_module
         return model
-
-    def transform(self) -> ExportedProgram:
-        ep = self.exported_program
-        for pass_ in self.passes:
-            signature = inspect.signature(pass_.__init__)
-            if "exported_program" in signature.parameters:
-                transform_pass = pass_(ep)
-            elif issubclass(pass_, ExportPass):
-                transform_pass = pass_()
-            else:
-                raise RuntimeError(
-                    f"Expecting ExportPass or ExportPass(), but got pass: {pass_} with type: {type(pass_)}"
-                )
-            ep = _transform(ep, transform_pass)
-        return ep
