+++ backends/cortex_m/quantizer/operator_configs.py
@@ -10,7 +10,6 @@ These can be used with the OperatorConfigQuantizer to quantize models based on o
 import torch
 
 from executorch.backends.cortex_m.quantizer.quantization_configs import (
-    INT8_PER_CHANNEL_CONFIG,
     INT8_PER_TENSOR_CONFIG,
 )
 from torchao.quantization.pt2e.quantizer import OperatorConfig
@@ -26,12 +25,6 @@ LINEAR_OP_PATTERNS = [
     [torch.ops.aten.linear.default, torch.ops.aten.relu.default],
 ]
 
-CONV_OP_PATTERNS = [
-    [torch.ops.aten.conv1d.default],
-    [torch.ops.aten.conv2d.default],
-    [torch.ops.aten.conv3d.default],
-]
-
 # ----------------- OPERATOR CONFIG PRESETS -----------------
 INT8_BINARY_OPS_OPERATOR_CONFIG = OperatorConfig(
     INT8_PER_TENSOR_CONFIG, BINARY_OP_PATTERNS
@@ -40,9 +33,4 @@ INT8_BINARY_OPS_OPERATOR_CONFIG = OperatorConfig(
 INT8_LINEAR_OPERATOR_CONFIG = OperatorConfig(
     INT8_PER_TENSOR_CONFIG,
     LINEAR_OP_PATTERNS,
-)
-
-INT8_CONV_OPERATOR_CONFIG = OperatorConfig(
-    INT8_PER_CHANNEL_CONFIG,
-    CONV_OP_PATTERNS,
 )
